+++ src/services/axios/put/putProposal/putProposal.tsx
@@ -7,7 +7,7 @@ import {
   USE_LOCAL_DATA
 } from '@utils/constants.ts';
 import Proposal, { ProposalBackend } from '@utils/types/proposal.tsx';
-import { AxiosAuthClient } from '../../axiosAuthClient/axiosAuthClient.tsx';
+import { AxiosAuthClient } from '../../axiosAuthClient/axiosAuthClient.ts';
 import MappingPutProposal from './putProposalMapping.tsx';
 import { MockProposalFrontend } from './mockProposalFrontend.tsx';